+++ crates/ldrs/tests/test_data/public.string_values/public.strings.snappy.parquet
